pyproject.toml
[tool.poetry]
name = "pyppium"
version = "0.1.0"
description = "Pyppium is a wrapper of Appium-Python-Client for cross mobile testing."
authors = ["Leonardo Menezes <leonardosmenezes.ssz@gmail.com>"]
repository = "https://github.com/leomenezessz/pyppium"
homepage = "https://github.com/leomenezessz/pyppium"
keywords = ['appium', 'qa', 'mobile-testing']
readme="README.md"
license = "MIT"

classifiers = [
    "Topic :: Utilities",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Libraries",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8"
]

packages = [
    { include = "pyppium" }
]

[tool.poetry.dependencies]
python = "^3.7"
Appium-Python-Client="==1.0.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
assertpy = "==1.0"
black = "==19.10b0"
pytest-mock = "==3.2.0"
pytest-cov= "==2.10.0"
codecov = "==2.1.7"
mkdocs = "==1.1.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
